[move] MinecraftServerThreadShift のパッケージを変更する
