+++ src/main/scala/com/github/unchama/menuinventory/Menu.scala
@@ -2,7 +2,7 @@ package com.github.unchama.menuinventory
 
 import cats.data
 import cats.effect.IO
-import com.github.unchama.concurrent.MinecraftServerThreadShift
+import com.github.unchama.minecraft.actions.MinecraftServerThreadShift
 import com.github.unchama.targetedeffect.TargetedEffect
 import org.bukkit.entity.Player
