+++ src/main/scala/com/github/unchama/minecraft/actions/SendMinecraftMessage.scala
@@ -1,6 +1,8 @@
 package com.github.unchama.minecraft.actions
 
-import cats.Applicative
+import cats.{Applicative, ~>}
+import com.github.unchama.generic.ContextCoercion
+import org.bukkit.entity.Player
 
 trait SendMinecraftMessage[F[_], Target] {
 
@@ -11,10 +13,17 @@ trait SendMinecraftMessage[F[_], Target] {
   def list(player: Target, strings: List[String])(implicit F: Applicative[F]): F[Unit] =
     strings.traverse(string(player, _)).as(())
 
+  def mapK[G[_]](fk: F ~> G): SendMinecraftMessage[G, Target] =
+    (player: Target, s: String) =>
+      fk(SendMinecraftMessage.this.string(player, s))
 }
 
 object SendMinecraftMessage {
 
   def apply[F[_], Player](implicit ev: SendMinecraftMessage[F, Player]): SendMinecraftMessage[F, Player] = implicitly
+
+  implicit def coercion[F[_], G[_]](implicit ev: SendMinecraftMessage[F, Player],
+                                    fg: ContextCoercion[F, G]): SendMinecraftMessage[G, Player] =
+    ev.mapK(fg)
 
 }
