+++ src/main/scala/com/github/unchama/seichiassist/subsystems/mebius/bukkit/repository/PeriodicMebiusSpeechRoutineFiberRepository.scala
@@ -1,8 +1,9 @@
 package com.github.unchama.seichiassist.subsystems.mebius.bukkit.repository
 
 import cats.effect.{Fiber, IO, SyncIO}
-import com.github.unchama.concurrent.{MinecraftServerThreadShift, RepeatingTaskContext}
+import com.github.unchama.concurrent.RepeatingTaskContext
 import com.github.unchama.datarepository.bukkit.player.JoinToQuitPlayerDataRepository
+import com.github.unchama.minecraft.actions.MinecraftServerThreadShift
 import com.github.unchama.seichiassist.subsystems.mebius.bukkit.routines.PeriodicMebiusSpeechRoutine
 import com.github.unchama.seichiassist.subsystems.mebius.service.MebiusSpeechService
 import org.bukkit.entity.Player
