+++ src/main/scala/com/github/unchama/seichiassist/subsystems/mebius/bukkit/routines/PeriodicMebiusSpeechRoutine.scala
@@ -2,8 +2,9 @@ package com.github.unchama.seichiassist.subsystems.mebius.bukkit.routines
 
 import cats.data.NonEmptyList
 import cats.effect.{IO, SyncIO, Timer}
-import com.github.unchama.concurrent.{MinecraftServerThreadShift, RepeatingRoutine, RepeatingTaskContext}
+import com.github.unchama.concurrent.{RepeatingRoutine, RepeatingTaskContext}
 import com.github.unchama.datarepository.bukkit.player.JoinToQuitPlayerDataRepository
+import com.github.unchama.minecraft.actions.MinecraftServerThreadShift
 import com.github.unchama.seichiassist.subsystems.mebius.bukkit.codec.BukkitMebiusItemStackCodec
 import com.github.unchama.seichiassist.subsystems.mebius.domain.resources.{MebiusMessages, MebiusTalks}
 import com.github.unchama.seichiassist.subsystems.mebius.domain.speech.{MebiusSpeech, MebiusSpeechStrength}
