+++ src/main/scala/com/github/unchama/seichiassist/task/global/PlayerDataRecalculationRoutine.scala
@@ -1,7 +1,8 @@
 package com.github.unchama.seichiassist.task.global
 
 import cats.effect.{IO, Timer}
-import com.github.unchama.concurrent.{MinecraftServerThreadShift, RepeatingRoutine, RepeatingTaskContext}
+import com.github.unchama.concurrent.{RepeatingRoutine, RepeatingTaskContext}
+import com.github.unchama.minecraft.actions.MinecraftServerThreadShift
 import com.github.unchama.seichiassist.SeichiAssist
 import com.github.unchama.seichiassist.achievement.SeichiAchievement
 import com.github.unchama.seichiassist.data.GachaSkullData
